Do not process events when firmware is not ready
This fixes some issues with error -28. Sometimes, the dongle is not
fully reset and it still has the dongle firmware. Sometimes it finds
nearby controllers before the driver finishes the whole init, loops the
add client event and runs out of possible client slots (16).

This change adds firmware state tracking to the dongle struct and
refuses to process any events if the firmware is not loaded. At least in
my testing, this made it so the client slots are never used and the
error goes away when firmware is loaded properly.

Fixes #35